fix attention data extraction to work with py37+
