+++ setup.cfg
@@ -10,7 +10,6 @@ license_files = LICENSE
 url = https://github.com/qismib/QuaKe.git
 classifiers =
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.9
     License :: OSI Approved :: MIT License
     Operating System :: Unix
     Topic :: Scientific/Engineering :: Physics
@@ -20,7 +19,7 @@ zip_safe = False
 package_dir=
     =src
 packages=find:
-python_requires = >=3.9
+python_requires = >=3.7
 install_requires =
     awkward
     numpy
@@ -30,7 +29,6 @@ install_requires =
     sklearn
     pyyaml
     uproot
-python_requires = >=3.7
 
 [options.packages.find]
 where = src
