+++ src/quake/dataset/generate_utils.py
@@ -118,18 +118,18 @@ def load_tracks(
         qtree = sig_root["qtree"]
 
         # (track, [hits]) nested branches
+        # uncomment the line below to introduce the random shift
+        # shouldn't z axis shift be reduced to uniform(-1,1)?
         normalize = (
-            lambda arr, arr0: arr
-            - arr0
-            + np.random.uniform(low=-5, high=5, size=1000)
-        )  # ak.mean(arr, axis=-1) arr[:,0]
+            lambda arr: arr # + np.random.uniform(low=-5, high=5, size=1000)
+        )
 
         xs = qtree["TrackPostX"].array()
         ys = qtree["TrackPostY"].array()
         zs = qtree["TrackPostZ"].array()
         Es = qtree["TrackEnergy"].array()
         tid = qtree["TrackID"].array()
-
+        
         if is_signal:
             # concatenate the two b tracks (from two consecutive rows)
             cat_fn = lambda arr: ak.concatenate([arr[::2], arr[1::2]], axis=1)
@@ -138,17 +138,16 @@ def load_tracks(
             zs = cat_fn(zs)
             Es = cat_fn(Es)
             tid = cat_fn(tid)
-        idx = list(np.sum(tid == 1, axis = 1)) # Index of the track starting point
-        xs0 = np.zeros(1000)
-        ys0 = np.zeros(1000)
-        zs0 = np.zeros(1000)
-        for i in range(0, 1000):
-            xs0 = xs[i, idx[i]]
-            ys0 = ys[i, idx[i]]
-            zs0 = zs[i, idx[i]]
-        xs = normalize(xs, xs0)
-        ys = normalize(ys, ys0)
-        zs = normalize(zs, zs0)
+            idx = np.sum(tid[:,::2] == 1, axis = 1) - 1 # Index of the track starting point
+        else:
+            idx = np.sum(tid == 1, axis = 1) - 1
+        idx = idx.to_numpy()
+        nev = idx.shape[0]
+        s_ = np.s_[np.arange(nev), idx]
+
+        xs = normalize(xs - xs[s_])
+        ys = normalize(ys - ys[s_])
+        zs = normalize(zs - zs[s_])
     return xs, ys, zs, Es
 
 
@@ -176,9 +175,9 @@ def tracks2histograms(
     hists = []
 
     # filter out of bin range hits
-    mx = np.logical_and(xs < geo.xbins[-1], xs >= geo.xbins[0])
-    my = np.logical_and(ys < geo.xbins[-1], ys >= geo.xbins[0])
-    mz = np.logical_and(zs < geo.xbins[-1], zs >= geo.xbins[0])
+    mx = np.logical_and(xs >= geo.xbins[0], xs < geo.xbins[-1])
+    my = np.logical_and(ys >= geo.xbins[0], ys < geo.xbins[-1])
+    mz = np.logical_and(zs >= geo.xbins[0], zs < geo.xbins[-1])
     m = np.logical_and(np.logical_and(mx, my), mz)
     for x, y, z, energy in zip(xs[m], ys[m], zs[m], Es[m]):
         # digits start from 1
@@ -192,6 +191,12 @@ def tracks2histograms(
         hist = sparse.csr_matrix(
             (energy.to_numpy(), (x_digits, yz_digits)), shape=shape
         )
+
+        try:
+            assert all(energy > 0)
+        except:
+            ValueError(f"Found non positive energies: {np.count_nonzero(energy > 0)}")
+
         hists.append(hist.reshape(1, -1))
     hists = sparse.vstack(hists)
     return hists
